requirements.txt
rasa==3.6.15
rasa-sdk==3.6.2
openai==0.28.0
python-dotenv
requests